refactor code

+++ user-service/src/main/java/com/stadyplanner/userservice/services/UserServiceImpl.java
@@ -57,8 +57,14 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(long userId) {
-        userRepository.deleteById(userId);
+    public boolean deleteUser(long userId) {
+        boolean exists = userRepository.existsById(userId);
+        if (!exists) {
+            return false;
+        } else {
+            userRepository.deleteById(userId);
+            return true;
+        }
     }
 
     @Override
